SHiELD-simulations/SHiELD/software/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile /gpfs/f5/gfdl_w/scratch/Spencer.Clark/projects/2024-01-23-SHiELD-SOM-simulations/environment/requirements.in
#
aiohttp==3.9.1
    # via gcsfs
aiosignal==1.3.1
    # via aiohttp
appdirs==1.4.4
    # via fv3config
attrs==23.2.0
    # via aiohttp
black==24.1.0
    # via -r /gpfs/f5/gfdl_w/scratch/Spencer.Clark/projects/2024-01-23-SHiELD-SOM-simulations/environment/requirements.in
cachetools==5.3.2
    # via google-auth
certifi==2023.11.17
    # via
    #   kubernetes
    #   requests
cftime==1.6.3
    # via -r /gpfs/f5/gfdl_w/scratch/Spencer.Clark/projects/2024-01-23-SHiELD-SOM-simulations/environment/requirements.in
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via black
dacite==1.8.1
    # via fv3config
decorator==5.1.1
    # via gcsfs
f90nml==1.4.4
    # via fv3config
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.12.2
    # via
    #   fv3config
    #   fv3kube
    #   gcsfs
fv3config==0.9.0
    # via fv3kube
fv3kube @ git+https://github.com/ai2cm/fv3net.git@68e4f378e933a9d90af9268a4f330cd6fc308e7d#subdirectory=external/fv3kube
    # via -r /gpfs/f5/gfdl_w/scratch/Spencer.Clark/projects/2024-01-23-SHiELD-SOM-simulations/environment/requirements.in
gcsfs==2023.12.2.post1
    # via fv3config
google-api-core==2.15.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.27.0
    # via
    #   gcsfs
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
    #   kubernetes
google-auth-oauthlib==1.2.0
    # via gcsfs
google-cloud-core==2.4.1
    # via google-cloud-storage
google-cloud-storage==2.14.0
    # via gcsfs
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.0
    # via google-cloud-storage
googleapis-common-protos==1.62.0
    # via google-api-core
idna==3.6
    # via
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
kubernetes==29.0.0
    # via fv3kube
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
numpy==1.26.3
    # via
    #   cftime
    #   pandas
    #   xarray
oauthlib==3.2.2
    # via
    #   kubernetes
    #   requests-oauthlib
packaging==23.2
    # via
    #   black
    #   pytest
    #   xarray
pandas==2.2.0
    # via xarray
pathspec==0.12.1
    # via black
platformdirs==4.1.0
    # via black
pluggy==1.4.0
    # via pytest
protobuf==4.25.2
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pytest==7.4.4
    # via -r /gpfs/f5/gfdl_w/scratch/Spencer.Clark/projects/2024-01-23-SHiELD-SOM-simulations/environment/requirements.in
python-dateutil==2.8.2
    # via
    #   kubernetes
    #   pandas
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   fv3config
    #   fv3kube
    #   kubernetes
requests==2.31.0
    # via
    #   gcsfs
    #   google-api-core
    #   google-cloud-storage
    #   kubernetes
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via
    #   google-auth-oauthlib
    #   kubernetes
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   kubernetes
    #   python-dateutil
tzdata==2023.4
    # via pandas
urllib3==2.1.0
    # via
    #   kubernetes
    #   requests
websocket-client==1.7.0
    # via kubernetes
xarray==2024.1.1
    # via -r /gpfs/f5/gfdl_w/scratch/Spencer.Clark/projects/2024-01-23-SHiELD-SOM-simulations/environment/requirements.in
yarl==1.9.4
    # via aiohttp
